pyproject.toml
[tool.poetry]
name = "token_talkers"
version = "0.1.0"
description = ""
authors = ["rochacbruno"]
readme = "README.md"
packages = [{include = "token_talkers"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.54.4"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
isort = "^5.13.2"
mypy = "^1.13.0"
gitchangelog = "^3.0.4"
mkdocs = "^1.6.1"
black = "^24.10.0"
coverage = "^7.6.4"
flake8 = "^7.1.1"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
token_talkers = 'token_talkers.__main__:main'

[tool.black]
line-length = 100
